requirements.txt
asgiref==3.8.1
Django==5.0.7
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==22.0.0
inflection==0.5.1
packaging==24.1
pytz==2024.1
PyYAML==6.0.1
setuptools==71.1.0
sqlparse==0.5.1
uritemplate==4.1.1
django-cors-headers==4.4.0
